Revert "Use int32 for positioning system internally"
This reverts commit 9109e75a650c88fce421700e8dc1f20565d7c929.

+++ src/core/core.go
@@ -225,8 +225,8 @@ func Run(newGameMode GameMode, enableProfiler bool) {
 
 			if MainCamera != nil {
 				RenderCamera.Offset = rl.Vector2{
-					X: float32(int(-float32(MainCamera.Position.X)*MainCamera.Zoom + float32(system.ScreenWidth)/2)),
-					Y: float32(int(-float32(MainCamera.Position.Y)*MainCamera.Zoom + float32(system.ScreenHeight)/2)),
+					X: float32(int(-MainCamera.Position.X*MainCamera.Zoom + float32(system.ScreenWidth)/2)),
+					Y: float32(int(-MainCamera.Position.Y*MainCamera.Zoom + float32(system.ScreenHeight)/2)),
 				}
 			}
